rename all file to be duo_universal_x.php
