+++ duo_universal_utilities.php
