+++ tests/duoUniversalAuthenticationTest.php
@@ -3,8 +3,8 @@
 use Duo\DuoUniversal\DuoException;
 use Duo\DuoUniversalWordpress;
 use PHPUnit\Framework\TestCase;
-require_once 'authentication.php';
-require_once 'duo_wordpress_helper.php';
+require_once 'duo_universal_authentication.php';
+require_once 'duo_universal_wordpress_helper.php';
 
 final class authenticationTest extends TestCase
 {
